frontend_bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.4.0
multidict==6.4.3
openai==1.76.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyTelegramBotAPI==4.26.0
python-docx==1.1.2
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
ffprobe
ffmpeg
aiofiles
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
sqlalchemy==2.0.27
asyncpg==0.29.0
minio==7.2.5
alembic==1.13.1